lib/trie: move trie from contrib to lib
